Squashed commit of the following:
commit ba7cff75995af7bd405ca4f8774d8c8ad352f27f
Merge: c210e66 b8fd2ee
Author: Zuul <zuul@lists.opencontrail.org>
Date:   Thu Jan 15 07:51:15 2015 -0800

    Merge "Partial-Bug: #1397496     Server Manager changes to make use of new structure puppet modules for contrail and openstack (from puppet-labs)."

commit b8fd2eec5bfe1258bbc9481f6d3a18de9d20426f
Author: tsurendra <tsurendra@juniper.net>
Date:   Tue Jan 13 11:19:41 2015 -0800

    Partial-Bug: #1397496
        Server Manager changes to make use of new structure puppet modules for contrail and openstack (from puppet-labs).

      Changes Done:
      1)Add status support
      2)Add additional params for rabbitmq and keepalived


commit c210e666bc8e0867c1d65af4e5b4b4316aba8b66
Author: tsurendra <tsurendra@juniper.net>
Date:   Thu Dec 11 19:54:25 2014 -0800

    [ubuntu14.04: upgrade kernel fails as package 'apparmor=2.7.102-0ubuntu3.10' was not available]

    Closes-Bug: #1401464

    Issue:
    kernel-upgrade for 14.04 doesnt require apparmour.
    as a result kernel upgrade fails.

    Fix:
    Bsaed on lsbrelease include Apparmour.

    Testing:
    1)Issued single node provision for ubuntu12.04 and ubuntu 14.04


commit d07e870cd80789354a2c65b18eb35ab426b45dad
Merge: df24dc8 0245277
Author: Zuul <zuul@lists.opencontrail.org>
Date:   Tue Jan 13 07:37:42 2015 -0800

    Merge "Added code to create same uid/gid for a set of users/groups."

commit 02452776144d40faa811657102c0def3828cb990
Author: Dheeraj Gautam <dgautam@juniper.net>
Date:   Tue Dec 16 03:22:08 2014 -0800

    Added code to create same uid/gid for a set of users/groups.

    created following groups
    'nova' =>  '499',
    'libvirtd' => '498',
    'kvm' => '497',

    and following users
    nova => 499,
    libvirt-qemu => 498,
    libvirt-dnsmasq => 497

    Unit Testing:
    =============
    provisioned two nodes with common and storage roles and verified that
    'ceph -s' is correct. Also verfied uid/gid configured on the system is
    correct.

    Partial-Bug: #1397496

+++ contrail/manifests/contrail_common.pp
@@ -34,20 +34,35 @@ define upgrade-kernel($contrail_kernel_version) {
 
     if ($operatingsystem == "Ubuntu")
     {
-        package { 'apparmor' : ensure => '2.7.102-0ubuntu3.10',}
-        ->
-        package { $headers : ensure => present, }
-        ->
-        package { $headers_generic : ensure => present, }
-        ->
-        package { $image : ensure => present, }
-        ->
-        exec { "upgrade-kernel-reboot":
-        command => "echo upgrade-kernel-reboot >> /etc/contrail/contrail_common_exec.out && reboot ",
-            provider => shell,
-            logoutput => "true",
-	    unless => ["grep -qx upgrade-kernel-reboot /etc/contrail/contrail_common_exec.out"]
-    	}
+        if ($lsbdistrelease == "14.04") {
+	    package { $headers : ensure => present, }
+	    ->
+	    package { $headers_generic : ensure => present, }
+	    ->
+	    package { $image : ensure => present, }
+	    ->
+	    exec { "upgrade-kernel-reboot":
+	    command => "echo upgrade-kernel-reboot >> /etc/contrail/contrail_common_exec.out && reboot ",
+		provider => shell,
+		logoutput => "true",
+		unless => ["grep -qx upgrade-kernel-reboot /etc/contrail/contrail_common_exec.out"]
+	    }
+        } else {
+	    package { 'apparmor' : ensure => '2.7.102-0ubuntu3.10',}
+	    ->
+	    package { $headers : ensure => present, }
+	    ->
+	    package { $headers_generic : ensure => present, }
+	    ->
+	    package { $image : ensure => present, }
+	    ->
+	    exec { "upgrade-kernel-reboot":
+	    command => "echo upgrade-kernel-reboot >> /etc/contrail/contrail_common_exec.out && reboot ",
+		provider => shell,
+		logoutput => "true",
+		unless => ["grep -qx upgrade-kernel-reboot /etc/contrail/contrail_common_exec.out"]
+	    }
+        }
      } else {
         #TODO for other flavours do nothing
      }
